fizzbuzz_server/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.2.0
blis==0.7.11
catalogue==2.0.10
celery==5.3.6
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.5.8
Django==4.2.2
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
docopt==0.6.2
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.4
langchain==0.0.201
langchainplus-sdk==0.0.10
langcodes==3.3.0
logging==0.4.9.6
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathlib==1.0.1
pinecone-client==2.2.2
pipreqs==0.4.13
preshed==3.0.9
prompt-toolkit==3.0.41
psycopg2==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==5.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.16
sqlparse==0.4.4
srsly==2.4.8
tenacity==8.2.2
textblob==0.17.1
thinc==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
twilio==8.3.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
win32-setctime==1.1.0
yarg==0.1.9
yarl==1.9.2
